Move UpdatePolicy into separate resource_attributes folder
