+++ lib/convection/model/template.rb
@@ -352,4 +352,5 @@ require_relative 'template/mapping'
 require_relative 'template/condition'
 require_relative 'template/resource'
 require_relative 'template/resource_property'
+require_relative 'template/resource_attribute'
 require_relative 'template/output'
