+++ lib/convection/model/template/resource/aws_auto_scaling_auto_scaling_group.rb
@@ -27,6 +27,8 @@ module Convection
           property :termination_policie, 'TerminationPolicies', :array
           property :vpc_zone_identifier, 'VPCZoneIdentifier', :array
 
+          property :network_interfaces, 'NetworkInterfaces', :type => :list
+
           def render(*args)
             super.tap do |resource|
               render_tags(resource)
@@ -35,47 +37,8 @@ module Convection
           end
 
           def update_policy(&block)
-            @update_policy = UpdatePolicy.new
-            @update_policy.instance_exec(&block)
-          end
-
-          ##
-          # UpdatePolicy for AWS::AutoScaling::AutoScalingGroup
-          ##
-          class UpdatePolicy
-            def initialize
-              @pause = 'PT5M'
-              @min_in_service = 0
-              @max_batch = 1
-            end
-
-            def pause_time(val)
-              @pause = val
-            end
-
-            def min_instances_in_service(val)
-              @min_in_service = val
-            end
-
-            def max_batch_size(val)
-              @max_batch = val
-            end
-
-            def render(resource)
-              resource.tap do |r|
-                r['UpdatePolicy'] = {
-                  'AutoScalingScheduledAction' => {
-                    'IgnoreUnmodifiedGroupSizeProperties' => true
-                  },
-                  'AutoScalingRollingUpdate' => {
-                    'MinInstancesInService' => @min_in_service,
-                    'MaxBatchSize' => @max_batch,
-                    'WaitOnResourceSignals' => false,
-                    'PauseTime' => @pause
-                  }
-                }
-              end
-            end
+            @update_policy = ResourceAttribute::UpdatePolicy.new(self)
+            @update_policy.instance_exec(&block) if block
           end
         end
       end
